Implemented Debt Instruments in Polars
