+++ streamlit/pages/1 - WTDOLLE - Stock Market Analysis.py
@@ -1,15 +1,12 @@
 import streamlit as st
-import pandas as pd
 import matplotlib.pyplot as plt
 
 from datetime import datetime
-from datetime import datetime
-import sys
 from functools import reduce
-
+import sys
 sys.path.append(".")
-from functions.generation_equities import Generate_DB, Generate_DB_polars, filter_indices
-from functions.generation_debt import Generate_Yield_panda
+
+from functions.generate_databases import Generate_Equity, Generate_Debt, filter_indices
 
 st.set_page_config(layout="wide",
                    page_title="What Transpired During Our Last Encounter (WTDOLLE)?",
@@ -54,7 +51,7 @@ nasdaq_intersection = []
 rus2k_intersection = []
 filtered_db_list=[] #to pass to filter_indices function
 
-sidebar_counter = 0
+equity_counter = 0
 st.subheader("")
 st.info(icon="ℹ️", body='SELECT METRICS HERE TO GENERATE EQUITY MARKET GRAPHS AND EVENTS. Note: Equity market metrics are available in production currently. Debt Market and Economic Figures are currently in progress.')
 inpcol1, inpcol2, inpcol3 = st.columns(3)
@@ -71,7 +68,7 @@ equityratio_check = equity_market.checkbox("Ratio of 2 Equities", False)
 # EQUITY MARKET -> VOLATILITY INDEX
 if volatility_index_check:
     with inpcol1.expander("Volatility Index"):
-        vix=Generate_DB_polars()
+        vix=Generate_Equity()
         vix.get_database(["^VIX"], input_start_date, input_end_date, input_interval)
         st.line_chart(vix.lf.select(["Date", "Close"]).collect(), x="Date", y="Close", height=200, use_container_width=True)
 
@@ -88,7 +85,7 @@ if volatility_index_check:
             filtered_db_list.append(vix.filtered_dates)
             
             equity_filters_applied_sentence+=f" VIX level {vix_level_comparator} {vix_level_selection}"
-            sidebar_counter+=1
+            equity_counter+=1
     #  EQUITY MARKET -> VOLATILITY INDEX -> VIX % CHANGE
         if vix_pct_on:
             vix_pct_lower = vixcol2.number_input("Between lower value", step=0.5, key="vix between lower value")
@@ -97,11 +94,11 @@ if volatility_index_check:
             vix.metric_vs_selection_cross(comparison_type="%_change",selected_value=[vix_pct_lower, vix_pct_higher], comparator="Between")
             filtered_db_list.append(vix.filtered_dates)
             
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" VIX % change between {vix_pct_lower}% and {vix_pct_higher}%"
             else:
                 equity_filters_applied_sentence+=f", VIX % change between {vix_pct_lower}% and {vix_pct_higher}%"
-            sidebar_counter+=1
+            equity_counter+=1
 
 # EQUITY MARKET -> RSP
 if equalweighted_sp500_check:
@@ -110,7 +107,7 @@ if equalweighted_sp500_check:
         rsp_rsi_length=rsp_chart_col1.number_input("Select RSI days", min_value=0, step=1, value=22, key="rsp rsi length selection")
         rsp_ma_length=rsp_chart_col2.number_input("Select MA days", min_value=0, step=1, value=50, key="rsp ma length selection")
 
-        rsp = Generate_DB_polars()
+        rsp = Generate_Equity()
         rsp.get_database('RSP', input_start_date, input_end_date, input_interval, rsi_value=rsp_rsi_length, ma_length=rsp_ma_length)
 
         st.line_chart(rsp.lf.select(["Date", "Close", "ma", "rsi"]).collect(), x="Date",y=["Close", "ma", "rsi"], height=200, use_container_width=True)
@@ -127,11 +124,11 @@ if equalweighted_sp500_check:
             rsp.metric_vs_selection_cross('rsi_vs_selection', [rsp_rsi_selection], rsp_rsi_comparator)
             filtered_db_list.append(rsp.filtered_dates)
             
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" RSP {rsp_rsi_length}-day RSI {rsp_rsi_comparator} {rsp_rsi_selection}"
             else:
                 equity_filters_applied_sentence+=f", RSP {rsp_rsi_length}-day RSI {rsp_rsi_comparator} {rsp_rsi_selection}"
-            sidebar_counter+=1
+            equity_counter+=1
         # EQUITY MARKET -> RSP -> RSP Moving Average
         if rsp_ma_on:
             rsp_ma_comparator = rsp_col2.selectbox(f"RSP Price > or < {rsp_ma_length} day Moving Average", ('Greater than', 'Less than'))
@@ -139,11 +136,11 @@ if equalweighted_sp500_check:
             rsp.metric_vs_selection_cross('price_vs_ma', rsp_ma_comparator)
             filtered_db_list.append(rsp.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" RSP Price {rsp_ma_comparator} RSP {rsp_ma_length} day Moving Average"
             else:
                 equity_filters_applied_sentence+=f", RSP Price {rsp_ma_comparator} RSP {rsp_ma_length} day Moving Average"
-            sidebar_counter+=1
+            equity_counter+=1
         
         # EQUITY MARKET -> RSP -> RSP RSI / Moving Average / % Change
         rsp_pct_on = rsp_col1.toggle("% Change", key="rsp % Change toggle")
@@ -156,11 +153,11 @@ if equalweighted_sp500_check:
             rsp.metric_vs_selection_cross("%_change","Between", [rsp_pct_lower, rsp_pct_higher])
             filtered_db_list.append(rsp.filtered_dates)
             
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" RSP % change between {rsp_pct_lower}% and {rsp_pct_higher}%"
             else:
                 equity_filters_applied_sentence+=f", RSP % change between {rsp_pct_lower} and {rsp_pct_higher}%"
-            sidebar_counter+=1
+            equity_counter+=1
 
 # EQUITY MARKET -> HYG
 if hyg_check:
@@ -169,7 +166,7 @@ if hyg_check:
         hyg_rsi_length=hyg_chart_col1.number_input("Select RSI days", min_value=0, step=1, value=22, key="hyg rsi length selection")
         hyg_ma_length=hyg_chart_col2.number_input("Select MA days", min_value=0, step=1, value=50, key="hyg ma length selection")
 
-        hyg = Generate_DB_polars()
+        hyg = Generate_Equity()
         hyg.get_database(['HYG'], input_start_date, input_end_date, input_interval, rsi_value=hyg_rsi_length, ma_length=hyg_ma_length)
         st.line_chart(hyg.lf.select(["Date", "Close", "ma", "rsi"]).collect(),x="Date",y=["Close", "ma", "rsi"], height=200, use_container_width=True)
 
@@ -186,11 +183,11 @@ if hyg_check:
             hyg.metric_vs_selection_cross("rsi_vs_selection", hyg_rsi_comparator, [hyg_rsi_selection])
             filtered_db_list.append(hyg.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" HYG {hyg_rsi_length}-day RSI {hyg_rsi_comparator} {hyg_rsi_selection}"
             else:
                 equity_filters_applied_sentence+=f", HYG {hyg_rsi_length}-day RSI {hyg_rsi_comparator} {hyg_rsi_selection}"
-            sidebar_counter+=1
+            equity_counter+=1
         # EQUITY MARKET -> HYG -> HYG Moving Average
         if hyg_ma_on:
             hyg_ma_comparator = hyg_col2.selectbox(f"HYG Price > or < {hyg_ma_length} day Moving Average", ('Greater than', 'Less than'))
@@ -198,11 +195,11 @@ if hyg_check:
             hyg.metric_vs_selection_cross("price_vs_ma", hyg_ma_comparator)
             filtered_db_list.append(hyg.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" HYG Price {hyg_ma_comparator} HYG {hyg_ma_length} day Moving Average"
             else:
                 equity_filters_applied_sentence+=f", HYG Price {hyg_ma_comparator} HYG {hyg_ma_length} day Moving Average"
-            sidebar_counter+=1
+            equity_counter+=1
         
         # EQUITY MARKET -> HYG -> HYG RSI / Moving Average / % Change
         hyg_pct_on = hyg_col1.toggle("% Change", key="hyg % Change toggle")
@@ -215,11 +212,11 @@ if hyg_check:
             hyg.metric_vs_selection_cross("%_change", "Between", [hyg_pct_lower, hyg_pct_higher])
             filtered_db_list.append(hyg.filtered_dates)
             
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" HYG % change between {hyg_pct_lower}% and {hyg_pct_higher}%"
             else:
                 equity_filters_applied_sentence+=f", HYG % change between {hyg_pct_lower} and {hyg_pct_higher}%"
-            sidebar_counter+=1
+            equity_counter+=1
 
 # EQUITY MARKET -> S&P 500
 if sp500_check:
@@ -230,7 +227,7 @@ if sp500_check:
         sp500_rsi_length=sp500_chartcol1.number_input("Select  RSI days", min_value=0, step=1, value=22, key="S&P500 rsi length selection")
         sp500_ma_length=sp500_chartcol2.number_input("Select MA days", min_value=0, step=1, value=50, key="S&P500 ma length selection")
         
-        sp500 = Generate_DB_polars()
+        sp500 = Generate_Equity()
         sp500.get_database(["^GSPC"], start_date=input_start_date, end_date=input_end_date, interval=input_interval, rsi_value=sp500_rsi_length, ma_length=sp500_ma_length)
         
         sp500_chartcol1.line_chart(sp500.lf.select(["rsi"]).collect(), height=200, use_container_width=True)
@@ -249,11 +246,11 @@ if sp500_check:
             sp500.metric_vs_selection_cross("rsi_vs_selection",sp500_rsi_comparator, [sp500_rsi_selection])
             filtered_db_list.append(sp500.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" S&P500 Price {sp500_rsi_comparator} {sp500_rsi_selection}"
             else:
                 equity_filters_applied_sentence+=f", S&P500 Price {sp500_rsi_comparator} {sp500_rsi_selection}"
-            sidebar_counter+=1
+            equity_counter+=1
 
         # EQUITY MARKET -> S&P 500 -> Moving Average
         if sp500_ma_on:
@@ -262,11 +259,11 @@ if sp500_check:
             sp500.metric_vs_selection_cross('price_vs_ma', sp500_ma_comparator)
             filtered_db_list.append(sp500.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" S&P500 Price {sp500_ma_comparator} S&P500 {sp500_ma_length} day Moving Average"
             else:
                 equity_filters_applied_sentence+=f", S&P500 Price {sp500_ma_comparator} S&P500 {sp500_ma_length} day Moving Average"
-            sidebar_counter+=1
+            equity_counter+=1
         
         # EQUITY MARKET -> S&P 500 -> % CHANGE / Price vs MA / RSI
         sp500_pct_on = sp500_col1.toggle("% Change", key="sp500 % Change toggle")
@@ -280,11 +277,11 @@ if sp500_check:
             
             sp500.metric_vs_selection_cross("%_change", "Between", [sp500_pct_lower, sp500_pct_higher])
             
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" S&P500 % change between {sp500_pct_lower}% and {sp500_pct_higher}%"
             else:
                 equity_filters_applied_sentence+=f", S&P500 % change between {sp500_pct_lower}% and {sp500_pct_higher}%"
-            sidebar_counter+=1
+            equity_counter+=1
             
 
 # EQUITY MARKET -> Nasdaq
@@ -295,7 +292,7 @@ if nasdaq_check:
         ndx_rsi_length=ndx_chartcol1.number_input("Select  RSI days", min_value=0, step=1, value=22, key="Nasdaq rsi length selection")
         ndx_ma_length=ndx_chartcol2.number_input("Select MA days", min_value=0, step=1, value=50, key="Nasdaq ma length selection")
         
-        ndx = Generate_DB_polars()
+        ndx = Generate_Equity()
         ndx.get_database("^IXIC", start_date=input_start_date, end_date=input_end_date, interval=input_interval, rsi_value=ndx_rsi_length, ma_length=ndx_ma_length)
         
         ndx_chartcol1.line_chart(ndx.lf.select(["Date", "rsi"]).collect(), x="Date", y="rsi", height=200, use_container_width=True)
@@ -314,11 +311,11 @@ if nasdaq_check:
             ndx.metric_vs_selection_cross("rsi_vs_selection", ndx_rsi_comparator, [ndx_rsi_selection])
             filtered_db_list.append(ndx.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" Nasdaq Price {ndx_rsi_comparator} {ndx_rsi_selection}"
             else:
                 equity_filters_applied_sentence+=f", Nasdaq Price {ndx_rsi_comparator} {ndx_rsi_selection}"
-            sidebar_counter+=1
+            equity_counter+=1
         # EQUITY MARKET -> Nasdaq -> Moving Average
         if ndx_ma_on:
             ndx_ma_comparator = ndx_col2.selectbox(f"ndx Price > or < {ndx_ma_length} day Moving Average", ('Greater than', 'Less than'))
@@ -326,11 +323,11 @@ if nasdaq_check:
             ndx.metric_vs_selection_cross("price_vs_ma", ndx_ma_comparator)
             filtered_db_list.append(ndx.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" Nasdaq Price {ndx_ma_comparator} Nasdaq {ndx_ma_length} day Moving Average"
             else:
                 equity_filters_applied_sentence+=f", Nasdaq Price {ndx_ma_comparator} Nasdaq {ndx_ma_length} day Moving Average"
-            sidebar_counter+=1
+            equity_counter+=1
         # EQUITY MARKET -> Nasdaq -> % CHANGE / Price vs MA / RSI
         ndx_pct_on = ndx_col1.toggle("% Change", key="ndx % Change toggle")
 
@@ -342,11 +339,11 @@ if nasdaq_check:
             ndx.metric_vs_selection_cross("%_change", "Between", [ndx_pct_lower, ndx_pct_higher])
             filtered_db_list.append(ndx.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" Nasdaq % change between {ndx_pct_lower}% and {ndx_pct_higher}%"
             else:
                 equity_filters_applied_sentence+=f", Nasdaq % change between {ndx_pct_lower}% and {ndx_pct_higher}%"
-            sidebar_counter+=1
+            equity_counter+=1
 
             ndx_col2.line_chart(ndx.lf.select(["Date", "% Change"]).collect(), x="Date", y="% Change", height=200, use_container_width=True)
 
@@ -358,7 +355,7 @@ if russell2000_check:
         rus2k_rsi_length=rus2k_chartcol1.number_input("Select  RSI days", min_value=0, step=1, value=22, key="Russell 2000 rsi length selection")
         rus2k_ma_length=rus2k_chartcol2.number_input("Select MA days", min_value=0, step=1, value=50, key="Russell 2000 ma length selection")
         
-        rus2k = Generate_DB_polars()
+        rus2k = Generate_Equity()
         rus2k.get_database("^RUT", start_date=input_start_date, end_date=input_end_date, interval=input_interval, rsi_value=rus2k_rsi_length, ma_length=rus2k_ma_length)
         
         rus2k_chartcol1.line_chart(rus2k.lf.select(["Date", "rsi"]).collect(), x="Date", y="rsi", height=200, use_container_width=True)
@@ -377,11 +374,11 @@ if russell2000_check:
             rus2k.metric_vs_selection_cross("rsi_vs_selection", rus2k_rsi_comparator, [rus2k_rsi_selection])
             filtered_db_list.append(rus2k.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" Russell 2000 Price {rus2k_rsi_comparator} {rus2k_rsi_selection}"
             else:
                 equity_filters_applied_sentence+=f", Russell 2000 Price {rus2k_rsi_comparator} {rus2k_rsi_selection}"
-            sidebar_counter+=1
+            equity_counter+=1
         # EQUITY MARKET -> Nasdaq -> Moving Average
         if rus2k_ma_on:
             rus2k_ma_comparator = rus2k_col2.selectbox(f"rus2k Price > or < {rus2k_ma_length} day Moving Average", ('Greater than', 'Less than'))
@@ -389,11 +386,11 @@ if russell2000_check:
             rus2k.metric_vs_selection_cross("price_vs_ma", rus2k_ma_comparator)
             filtered_db_list.append(rus2k.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" Russell 2000 Price {rus2k_ma_comparator} Russell 2000 {rus2k_ma_length} day Moving Average"
             else:
                 equity_filters_applied_sentence+=f", Russell 2000 Price {rus2k_ma_comparator} Russell 2000 {rus2k_ma_length} day Moving Average"
-            sidebar_counter+=1
+            equity_counter+=1
         # EQUITY MARKET -> Russell 2000 -> % CHANGE / Price vs MA / RSI
         rus2k_pct_on = rus2k_col1.toggle("% Change", key="rus2k % Change toggle")
 
@@ -405,11 +402,11 @@ if russell2000_check:
             rus2k.metric_vs_selection_cross("%_change", "Between", [rus2k_pct_lower, rus2k_pct_higher])
             filtered_db_list.append(rus2k.filtered_dates)
 
-            if sidebar_counter==0:
+            if equity_counter==0:
                 equity_filters_applied_sentence+=f" Russell 2000 % change between {rus2k_pct_lower}% and {rus2k_pct_higher}%"
             else:
                 equity_filters_applied_sentence+=f", Russell 2000 % change between {rus2k_pct_lower}% and {rus2k_pct_higher}%"
-            sidebar_counter+=1
+            equity_counter+=1
 
             rus2k_col2.line_chart(rus2k.lf.select(["Date", "% Change"]).collect(),x="Date", y="% Change", height=200, use_container_width=True)
 
@@ -424,7 +421,7 @@ if equityratio_check:
             eq_ratio_rsi_length = eq_ratio_col1.number_input("Select RSI days", min_value=0, step=1, value=22, key="equity ratio RSI length")
             eq_ratio_ma_length = eq_ratio_col2.number_input("Select MA days", min_value=0, step=1, value=50, key="equity ratio MA length")
             
-            equity_ratio = Generate_DB_polars()
+            equity_ratio = Generate_Equity()
             equity_ratio.generate_ratio(eq_ratio_numerator_selection, eq_ratio_denominator_selection, input_start_date, input_end_date, input_interval, eq_ratio_rsi_length, eq_ratio_ma_length)
             
             eq_ratio_col1.line_chart(equity_ratio.lf.select(["Date", "rsi"]).collect(), x="Date", y=["Date", "rsi"], height=200, use_container_width=True)
@@ -442,11 +439,11 @@ if equityratio_check:
                 equity_ratio.metric_vs_selection_cross("rsi_vs_selection", eq_ratio_comparator, [eq_ratio_rsi_selection])
                 filtered_db_list.append(equity_ratio.filtered_dates)
 
-                if sidebar_counter==0:
+                if equity_counter==0:
                     equity_filters_applied_sentence+=f" Equity Ratio Price {eq_ratio_comparator} {eq_ratio_rsi_selection}"
                 else:
                     equity_filters_applied_sentence+=f", Equity Ratio Price {eq_ratio_comparator} {eq_ratio_rsi_selection}"
-                sidebar_counter+=1
+                equity_counter+=1
             
             # EQUITY MARKET -> EQUITY RATIO -> MA
             if eq_ratio_ma_on:
@@ -455,11 +452,11 @@ if equityratio_check:
                 equity_ratio.metric_vs_selection_cross("price_vs_ma", eq_ratio_ma_comparator)
                 filtered_db_list.append(equity_ratio.filtered_dates)                
                 
-                if sidebar_counter==0:
+                if equity_counter==0:
                     equity_filters_applied_sentence+=f" Equity Ratio Price {eq_ratio_ma_comparator} Equity Ratio {eq_ratio_ma_length} day Moving Average"
                 else:
                     equity_filters_applied_sentence+=f", Equity Ratio Price {eq_ratio_ma_comparator} Equity Ratio {eq_ratio_ma_length} day Moving Average"
-                sidebar_counter+=1
+                equity_counter+=1
             
             # EQUITY MARKET -> EQUITY RATIO -> % CHANGE
             eq_ratio_pct_on=eq_ratio_col1.toggle("% Change", key="equity ratio % change")
@@ -472,11 +469,11 @@ if equityratio_check:
                 equity_ratio.metric_vs_selection_cross("%_change", "Between", [eqratio_pct_lower, eqratio_pct_higher])
                 filtered_db_list.append(equity_ratio.filtered_dates)
 
-                if sidebar_counter==0:
+                if equity_counter==0:
                     equity_filters_applied_sentence+=f" Equity Ratio % change between {eqratio_pct_lower}% and {eqratio_pct_higher}%"
                 else:
                     equity_filters_applied_sentence+=f", Equity Ratio % change between {eqratio_pct_lower}% and {eqratio_pct_higher}%"
-                sidebar_counter+=1
+                equity_counter+=1
 
 # EQUITY MARKET -> SUMMARY
 inpcol1.write("*"+equity_filters_applied_sentence+"*")
@@ -486,12 +483,13 @@ debt_filters_applied_sentence = "Debt filters applied:"
 debt_market = inpcol2.popover("Debt Market")
 yieldspread_check = debt_market.checkbox("Market Yield Spread (Yield Curve)", False)
 usfedfundrate_check = debt_market.checkbox("US Federal Funds Rate", False)
+debt_counter=0
 
 
 # DEBT MARKET -> YIELD SPREAD
 if yieldspread_check:
     with inpcol2.expander("Yield Spread"):
-        yieldspread = Generate_Yield_panda()
+        yieldspread = Generate_Debt()
         spreadcol1, spreadcol2 = st.columns(2)
 
         # DEBT MARKET -> YIELD SPREAD -> Long Term
@@ -501,13 +499,13 @@ if yieldspread_check:
         st_maturity_selection = spreadcol2.selectbox("Short-Term Maturity", ("1m", "3m", "6m", "1y", "2y", "3y", "5y", "7y", "10y", "20y", "30y"), index=4)
 
         # DEBT MARKET -> YIELD SPREAD
-        yieldspread.generate_yield_spread(input_start_date, input_end_date, selection_interval, lt_maturity_selection, st_maturity_selection)
-        spread_linechart = yieldspread.yielddiff_df[[f'{lt_maturity_selection} Rate', f'{st_maturity_selection} Rate', 'Rate Spread', '% Change']]
-        spreadcol1.line_chart(spread_linechart, height=200, use_container_width=True, y=[f'{lt_maturity_selection} Rate', f'{st_maturity_selection} Rate', 'Rate Spread'], color=['#c9c9e6', '#cce6c9', '#be2a25'])
-        spreadcol2.dataframe(spread_linechart, height=207, column_order=['Rate Spread', '% Change'],column_config={
-            "": st.column_config.DatetimeColumn(
-                "Date",
-                format="YYYY-MM-DD"
+        yieldspread.generate_yield_spread(input_start_date, input_end_date, lt_maturity_selection, st_maturity_selection, selection_interval)
+
+        spreadcol1.line_chart(yieldspread.lf.select(["Date", f'{lt_maturity_selection} Rate', f'{st_maturity_selection} Rate', 'Rate Spread']).collect(), height=200, use_container_width=True,x="Date", y=[f'{lt_maturity_selection} Rate', f'{st_maturity_selection} Rate', 'Rate Spread'], color=['#c9c9e6', '#cce6c9', '#be2a25'])
+
+        spreadcol2.dataframe(yieldspread.lf.select(["Date", 'Rate Spread', "% Change"]).collect(), height=207, hide_index=True,column_order=["Date",'Rate Spread', '% Change'],column_config={
+            "Date": st.column_config.DateColumn(
+                "Date", format="YYYY-MM-DD"
             ),
             '% Change': st.column_config.NumberColumn(
                 "% Change",
@@ -522,44 +520,53 @@ if yieldspread_check:
         if spread_level_on:
             spread_level_comparator = spreadcol1.selectbox("Spread Comparator", ('Greater than', 'Less than'))
             spread_level_selection=spreadcol1.number_input("Select value", step=1.0, value=0.0)
-            sp500_intersection, nasdaq_intersection, rus2k_intersection = yieldspread.metric_vs_selection(movement='cross',comparison_type='current price', comparator=spread_level_comparator, selected_value=[spread_level_selection], sp500=sp500_intersection, ndx=nasdaq_intersection, rus2k=rus2k_intersection)
-            
-            debt_filters_applied_sentence+=f" Spread Level {spread_level_comparator} {spread_level_selection}"
-            sidebar_counter+=1
+
+            yieldspread.metric_vs_selection_cross("yield_spread", spread_level_comparator, [spread_level_selection])
+            filtered_db_list.append(yieldspread.filtered_dates)
+
+            if debt_counter == 0:            
+                debt_filters_applied_sentence+=f" Spread Level {spread_level_comparator} {spread_level_selection}"
+            else:
+                debt_filters_applied_sentence+=f", Spread Level {spread_level_comparator} {spread_level_selection}"
+            debt_counter+=1
         
         if spread_pct_on:
             spread_pct_lower = spreadcol2.number_input("Between lower value", step=0.5, key="spread between lower value")
             spread_pct_higher = spreadcol2.number_input("Between higher value", step=0.5, key="spread between higher value")
-            sp500_intersection, nasdaq_intersection, rus2k_intersection = yieldspread.metric_vs_selection(movement='cross', comparison_type='% change between', comparator="Between", selected_value=[spread_pct_lower, spread_pct_higher], sp500=sp500_intersection, ndx=nasdaq_intersection, rus2k=rus2k_intersection)
 
-            if sidebar_counter==0:
+            yieldspread.metric_vs_selection_cross("%_change", "Between", [spread_pct_lower, spread_pct_higher])
+            filtered_db_list.append(yieldspread.filtered_dates)
+
+            if debt_counter==0:
                 debt_filters_applied_sentence+=f" Spread % change between {spread_pct_lower}% and {spread_pct_higher}%"
             else:
                 debt_filters_applied_sentence+=f", Spread % change between {spread_pct_lower}% and {spread_pct_higher}%"
-            sidebar_counter+=1
+            debt_counter+=1
 
 if usfedfundrate_check:    
     with inpcol2.expander("US Federal Funds Rate"):
-        usfedfund = Generate_Yield_panda()
+        usfedfund = Generate_Debt()
         usfedfund.get_database("US FED FUNDS", input_start_date, input_end_date, selection_interval)
         
         usfedr_col1, usdfedr_col2 = st.columns(2)
 
         # DEBT MARKET -> US FED FUND RATE -> Change
-        usfedr_col1.line_chart(usfedfund.df, y=['US FED FUNDS Rate'], height=200)
+        usfedr_col1.line_chart(usfedfund.lf.collect(),x="Date", y=['US FED FUNDS Rate'], height=200)
         
         usfedrate_level_on = usdfedr_col2.toggle("US Fed Funds Rate Change", key="US Fed Funds Rate Change")
 
         if usfedrate_level_on:
             usfedrate_level_selection_lower = usdfedr_col2.number_input("Increase/Decrease Between (lower value)", step=0.25, value=-0.30)
             usfedrate_level_selection_higher = usdfedr_col2.number_input("Increase/Decrease Between (higher value)", step=0.25, value=-0.20)
-            sp500_intersection, nasdaq_intersection, rus2k_intersection = usfedfund.metric_vs_selection(movement='cross', comparison_type='change between', comparator="change between", selected_value=[usfedrate_level_selection_lower, usfedrate_level_selection_higher], sp500=sp500_intersection, ndx=nasdaq_intersection, rus2k=rus2k_intersection)
 
-            if sidebar_counter==0:
+            usfedfund.metric_vs_selection_cross("%_change", "Between", [usfedrate_level_selection_lower, usfedrate_level_selection_higher])
+            filtered_db_list.append(usfedfund.filtered_dates)
+
+            if debt_counter==0:
                 debt_filters_applied_sentence+=f" US Fed Funds Rate change between {usfedrate_level_selection_lower}% and {usfedrate_level_selection_higher}%"
             else:
                 debt_filters_applied_sentence+=f", US Fed Funds Rate change between {usfedrate_level_selection_lower}% and {usfedrate_level_selection_higher}%"
-            sidebar_counter+=1
+            debt_counter+=1
 
 # DEBT MARKET -> SUMMARY
 inpcol2.write("*"+debt_filters_applied_sentence+"*")
@@ -567,92 +574,95 @@ inpcol2.write("*"+debt_filters_applied_sentence+"*")
 inpcol3.subheader("Economic Figures - IN PROGRESS")
 
 
-# --- SELECTED VARIABLES COLUMNS ---
-col1, col2, col3, col4, col5, col6, col7, col8 = st.columns(8)
+st.button("APPLY FILTERS", use_container_width=True, type="primary")
+
+def apply_filters():
+    # --- MAIN PAGE OF WTDOLLE - CHARTS AND PARAMETERS
+        # -------------------- HEADER -------------------------
+    st.header(f'WTDOLLE on {input_end_date}')
+
+    # ---------MODIFYING THE METRIC FORMAT ------------------
+    st.markdown(
+        """
+    <style>
+    [data-testid="stMetricValue"] {
+        font-size: 20px;
+    }
+    </style>
+    """,
+        unsafe_allow_html=True,
+    )
+
+    #--- PLOTTING GRAPH ---
+    col1, col2, col3 = st.columns(3)
+    graph1, graph2, graph3= st.columns(3)
+
+    #-------INDICES PARAMETER SELECTION-------
+    col1.subheader("S&P 500")
+
+    col2.subheader("Nasdaq")
+
+    col3.subheader("Russell 2000")
+
+    # --- Indices Generation ---
+    if equity_counter+debt_counter > 0:
+        #---S&P500 DATABASE GENERATION---
+        if 'sp500' not in globals():
+            sp500 = Generate_DB()
+            sp500.get_database("^GSPC", input_start_date, input_end_date, input_interval)
+        sp500.generate_common_dates(sp500_intersection,selected_returninterval=input_returninterval)
+        #-------S&P500 GRAPH------
+        fig, ax = plt.subplots()
+        ax.set_title('S&P500')
+        ax.plot(sp500.db.index, sp500.db['Close'], linewidth = 0.5, color='black')
+        try:
+            ax.scatter(sp500.common_dates.index, sp500.common_dates['Close'], marker='.', color='red', s = 10)
+            graph1.pyplot(fig)
+            #-------S&P500 GENERATE STATEMENTS--------
+            graph1.write(f'This occurred {sp500.no_of_occurrences} of time(s) and is {sp500.positive_percentage} positive in {input_returninterval} days.' )
+            graph1.write('{:.2%}'.format(sp500.avg_return))
+        except AttributeError:
+            graph1.write("There are no scenarios that exist like this.")
 
 
-# --- MAIN PAGE OF WTDOLLE - CHARTS AND PARAMETERS
-    # -------------------- HEADER -------------------------
-st.header(f'WTDOLLE on {input_end_date}')
 
-# ---------MODIFYING THE METRIC FORMAT ------------------
-st.markdown(
-    """
-<style>
-[data-testid="stMetricValue"] {
-    font-size: 20px;
-}
-</style>
-""",
-    unsafe_allow_html=True,
-)
+        #---NASDAQ DATABASE GENERATION---
+        if 'ndx' not in globals():
+            ndx = Generate_DB()
+            ndx.get_database("^IXIC", input_start_date, input_end_date, input_interval)
+        ndx.generate_common_dates(nasdaq_intersection, selected_returninterval=input_returninterval)
+        #-------NASDAQ GRAPH------
+        fig, ax = plt.subplots()
+        ax.set_title('Nasdaq 100')
+        try:
+            ax.plot(ndx.db.index, ndx.db['Close'], linewidth = 0.5, color='black')
+            ax.scatter(ndx.common_dates.index, ndx.common_dates['Close'], marker='.', color='red', s = 10)
+            graph2.pyplot(fig)
+            #-------NASDAQ GENERATE STATEMENTS-------
+            graph2.write(f'This occurred {ndx.no_of_occurrences} time(s) and is {ndx.positive_percentage} positive in {input_returninterval} days.' )
+            graph2.write('{:.2%}'.format(ndx.avg_return))
+        except AttributeError:
+            graph2.write("There are no scenarios that exist like this.")
+        
 
-#--- PLOTTING GRAPH ---
-col1, col2, col3 = st.columns(3)
-graph1, graph2, graph3= st.columns(3)
 
-#-------INDICES PARAMETER SELECTION-------
-col1.subheader("S&P 500")
+        #--RUSSEL 2000 DATABASE GENERATION
+        if 'rus2k' not in globals():
+            rus2k = Generate_DB()
+            rus2k.get_database("^RUT", input_start_date, input_end_date, input_interval)
+        rus2k.generate_common_dates(rus2k_intersection, selected_returninterval=input_returninterval)
+        #-----RUSSELL 2000 GRAPH-----
+        fig, ax = plt.subplots()
+        ax.set_title('Russel 2000')
+        try:
+            ax.plot(rus2k.db.index, rus2k.db['Close'], linewidth = 0.5, color='black')
+            ax.scatter(rus2k.common_dates.index, rus2k.common_dates['Close'], marker='.', color='red', s = 10)
+            graph3.pyplot(fig)
+            #-------NASDAQ GENERATE STATEMENTS-------
+            graph3.write(f'This occurred {rus2k.no_of_occurrences} of time(s) and is {rus2k.positive_percentage} positive in {input_returninterval} days.' )
+            graph3.write('{:.2%}'.format(rus2k.avg_return))
+        except AttributeError:
+            graph3.write("There are no scenarios that exist like this.")
 
-col2.subheader("Nasdaq")
-
-col3.subheader("Russell 2000")
-
-# --- Indices Generation ---
-if sidebar_counter > 0:
-    #---S&P500 DATABASE GENERATION---
-    if 'sp500' not in globals():
-        sp500 = Generate_DB()
-        sp500.get_database("^GSPC", input_start_date, input_end_date, input_interval)
-    sp500.generate_common_dates(sp500_intersection,selected_returninterval=input_returninterval)
-    #-------S&P500 GRAPH------
-    fig, ax = plt.subplots()
-    ax.set_title('S&P500')
-    ax.plot(sp500.db.index, sp500.db['Close'], linewidth = 0.5, color='black')
-    try:
-        ax.scatter(sp500.common_dates.index, sp500.common_dates['Close'], marker='.', color='red', s = 10)
-        graph1.pyplot(fig)
-        #-------S&P500 GENERATE STATEMENTS--------
-        graph1.write(f'This occurred {sp500.no_of_occurrences} of time(s) and is {sp500.positive_percentage} positive in {input_returninterval} days.' )
-        graph1.write('{:.2%}'.format(sp500.avg_return))
-    except AttributeError:
-        graph1.write("There are no scenarios that exist like this.")
-
-    #---NASDAQ DATABASE GENERATION---
-    if 'ndx' not in globals():
-        ndx = Generate_DB()
-        ndx.get_database("^IXIC", input_start_date, input_end_date, input_interval)
-    ndx.generate_common_dates(nasdaq_intersection, selected_returninterval=input_returninterval)
-    #-------NASDAQ GRAPH------
-    fig, ax = plt.subplots()
-    ax.set_title('Nasdaq 100')
-    try:
-        ax.plot(ndx.db.index, ndx.db['Close'], linewidth = 0.5, color='black')
-        ax.scatter(ndx.common_dates.index, ndx.common_dates['Close'], marker='.', color='red', s = 10)
-        graph2.pyplot(fig)
-        #-------NASDAQ GENERATE STATEMENTS-------
-        graph2.write(f'This occurred {ndx.no_of_occurrences} time(s) and is {ndx.positive_percentage} positive in {input_returninterval} days.' )
-        graph2.write('{:.2%}'.format(ndx.avg_return))
-    except AttributeError:
-        graph2.write("There are no scenarios that exist like this.")
-    
-    #--RUSSEL 2000 DATABASE GENERATION
-    if 'rus2k' not in globals():
-        rus2k = Generate_DB()
-        rus2k.get_database("^RUT", input_start_date, input_end_date, input_interval)
-    rus2k.generate_common_dates(rus2k_intersection, selected_returninterval=input_returninterval)
-    #-----RUSSELL 2000 GRAPH-----
-    fig, ax = plt.subplots()
-    ax.set_title('Russel 2000')
-    try:
-        ax.plot(rus2k.db.index, rus2k.db['Close'], linewidth = 0.5, color='black')
-        ax.scatter(rus2k.common_dates.index, rus2k.common_dates['Close'], marker='.', color='red', s = 10)
-        graph3.pyplot(fig)
-        #-------NASDAQ GENERATE STATEMENTS-------
-        graph3.write(f'This occurred {rus2k.no_of_occurrences} of time(s) and is {rus2k.positive_percentage} positive in {input_returninterval} days.' )
-        graph3.write('{:.2%}'.format(rus2k.avg_return))
-    except AttributeError:
-        graph3.write("There are no scenarios that exist like this.")
-
-else:
-    pass
+    else:
+        pass
